requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0